annotation_table - DOC/BUG - fixed bug in create_rndm_backgr and added docstring example
Former-commit-id: 0087a3885b3a209deb89748cb134b6cda55228eb
Former-commit-id: 23ebace739692aa55d6acb07372b46c1428408e1
Former-commit-id: 49f94bf7b40f10e54310baf5a893a8638403477d
Former-commit-id: 44200c4b6e0f081ba0c3b68d284c5fc1e29b55e8

+++ ketos/tests/conftest.py
@@ -336,7 +336,7 @@ def annot_table_std():
     label = [1, 2, 3, 0, 0, -1]
     N = len(label)
     filename = ['f{0}.wav'.format(x%3) for x in np.arange(N)]
-    start = np.arange(N)
+    start = np.arange(N, dtype=float)
     stop = start + 3.3
     tbl = pd.DataFrame({'filename': filename, 'label': label, 'time_start': start, 'time_stop': stop})
     return tbl
